Packages: Extract @wordpress/scripts-utils package
