+++ packages/scripts-utils/README.md
@@ -1,0 +1,28 @@
+# Scripts Utils
+
+Collection of scripts utils for WordPress development.
+
+## Installation
+
+You only need to install one npm module:
+
+```bash
+npm install @wordpress/scripts-utils --save-dev
+```
+
+## API
+
+**Methods included**
+
+- `fromProjectRoot`
+- `getArgFromCLI`
+- `getArgsFromCLI`
+- `getFileArgsFromCLI`
+- `hasArgInCLI`
+- `hasFileArgInCLI`
+- `hasPackageProp`
+- `hasProjectFile`
+
+**Note**: This package requires `node` 10.0.0 or later. It is not compatible with older versions.
+
+<br/><br/><p align="center"><img src="https://s.w.org/style/images/codeispoetry.png?1" alt="Code is Poetry." /></p>
