+++ packages/scripts-utils/lib/package.js
