+++ packages/scripts-utils/lib/process.js
@@ -1,0 +1,16 @@
+const getArgsFromCLI = ( excludePrefixes ) => {
+	const args = process.argv.slice( 2 );
+	if ( excludePrefixes ) {
+		return args.filter( ( arg ) => {
+			return ! excludePrefixes.some( ( prefix ) =>
+				arg.startsWith( prefix )
+			);
+		} );
+	}
+	return args;
+};
+
+module.exports = {
+	getArgsFromCLI,
+	getCurrentWorkingDirectory: process.cwd,
+};
