+++ packages/scripts/scripts/lint-md-js.js
@@ -5,15 +5,19 @@ const { sync: spawn } = require( 'cross-spawn' );
 const { sync: resolveBin } = require( 'resolve-bin' );
 
 /**
- * Internal dependencies
+ * WordPress dependencies
  */
 const {
-	fromConfigRoot,
 	getArgsFromCLI,
 	hasArgInCLI,
 	hasFileArgInCLI,
 	hasProjectFile,
-} = require( '../utils' );
+} = require( '@wordpress/scripts-utils' );
+
+/**
+ * Internal dependencies
+ */
+const { fromConfigRoot } = require( '../utils' );
 
 const args = getArgsFromCLI();
