+++ packages/scripts/utils/file.js
@@ -4,17 +4,6 @@
 const { existsSync, readdirSync } = require( 'fs' );
 const path = require( 'path' );
 
-/**
- * Internal dependencies
- */
-const { getPackagePath } = require( './package' );
-
-const fromProjectRoot = ( fileName ) =>
-	path.join( path.dirname( getPackagePath() ), fileName );
-
-const hasProjectFile = ( fileName ) =>
-	existsSync( fromProjectRoot( fileName ) );
-
 const fromConfigRoot = ( fileName ) =>
 	path.join( path.dirname( __dirname ), 'config', fileName );
 
@@ -30,10 +19,8 @@ const getScripts = () =>
 		.map( ( f ) => path.basename( f, '.js' ) );
 
 module.exports = {
-	fromProjectRoot,
 	fromConfigRoot,
 	fromScriptsRoot,
 	getScripts,
-	hasProjectFile,
 	hasScriptFile,
 };
